+++ Colors.xcassets/Color6.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0x00",
-          "green" : "0xAA",
-          "red" : "0x99"
+          "blue" : "0x23",
+          "green" : "0xB7",
+          "red" : "0x52"
         }
       },
       "idiom" : "universal"
@@ -23,9 +23,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0x00",
-          "green" : "0xAA",
-          "red" : "0x99"
+          "blue" : "0x23",
+          "green" : "0xB7",
+          "red" : "0x52"
         }
       },
       "idiom" : "universal"
