+++ Rephoto_iOS/Features/Search/Models/SearchModel.swift
